rust: Deal with new Rust 1.66 warnings (#977)
* CI: Correct Rust toolchain installation

Not only install the toolchain specified by the matrix, but also make
it the default one, since the build system uses the default toolchain.

Currently, our "1.51" build actually uses "stable" because of this.

:facepalm:

* rust: Minimum supported version is now 1.56

Guess what? CI was lying. RustThemis does not compile with 1.51 because
we use Rust 2021 edition which was not stabilized until 1.56. Bump it.

* rust: Drop unnecessary casts

New stable Clippy points out the following:

     error: casting to the same type is unnecessary (`usize` -> `usize`)
       --> src/wrappers/themis/rust/src/keygen.rs:108:29
        |
    108 |         private_key.set_len(private_key_len as usize);
        |                             ^^^^^^^^^^^^^^^^^^^^^^^^ help: try: `private_key_len`
        |
        = help: for further information visit https://rust-lang.github.io/rust-clippy/master/index.html#unnecessary_cast
        = note: `-D clippy::unnecessary-cast` implied by `-D warnings`

* rust: Suppress more warnings

Previously this was suppressed in the main Themis code, but seems like
the new Clippy learned about the existence of benchmarks. Suppress this
warning there as well.

+++ docs/examples/rust/secure_session_echo_server.rs
@@ -78,7 +78,7 @@ impl SecureSessionTransport for SocketTransport {
         }
         self.socket.read_exact(&mut data[0..length])?;
         debug!("{:?}: received {} bytes", self.socket.peer_addr(), length);
-        Ok(length as usize)
+        Ok(length)
     }
 
     fn state_changed(&mut self, state: SecureSessionState) {
